+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- CLI new-config-file handle error when directory does not exist.
+
 ## [0.3.0] - 2024-03-19
 ### Adding
 - Allow editing configs for a given block